Bug fix: didn't give result descriptor type in extractRoubstFeature(). Add test on regression matching distance threshold.

+++ src/main/java/edu/umb/cs/imageprocessinglib/feature/FeatureMatcher.java
@@ -143,7 +143,7 @@ public class FeatureMatcher {
                                            MatOfKeyPoint templateKeyPoints,
                                            DescriptorType dType,
                                            int knnNum,
-                                           int matchDisThd,
+                                           float matchDisThd,
                                            int posThd) {
 //        MatOfDMatch matches = new MatOfDMatch();
 //        matcher.match(queryDescriptor, templateDescriptor, matches);       //k(final parameter) set to 1 will do crosscheck
